These changes update the various profiles and methods in Characterization as a result of some recent changes in ds-java master.

+++ src/main/java/org/apache/datasketches/MonotonicPoints.java
@@ -150,8 +150,8 @@ public class MonotonicPoints {
    * @return the actual number of plotting points between lgStart and lgEnd.
    */
   public static final int countPoints(final int lgStart, final int lgEnd, final int ppo) {
-    int p = 1 << lgStart;
-    final int end = 1 << lgEnd;
+    long p = 1L << lgStart;
+    final long end = 1L << lgEnd;
     int count = 0;
     while (p <= end) {
       p = pwr2SeriesNext(ppo, p);
