+++ src/main/java/org/apache/datasketches/characterization/memory/BaseSpeedProfile.java
@@ -117,7 +117,7 @@ static class Point {
     final int minX = 1 << lgMinX;
     int lastX = 0;
     while (lastX < maxX) {
-      final int nextX = lastX == 0 ? minX : pwr2SeriesNext(xPPO, lastX);
+      final int nextX = lastX == 0 ? minX : (int)pwr2SeriesNext(xPPO, lastX);
       lastX = nextX;
       final int trials = getNumTrials(nextX);
       configure(nextX);
