+++ src/main/java/org/apache/datasketches/characterization/memory/BaseUtf8SpeedProfile.java
@@ -154,7 +154,7 @@ public abstract class BaseUtf8SpeedProfile implements JobProfile {
     final int minX = 1 << lgMinX;
     int lastX = 0;
     while (lastX < maxX) { //do each plot point on the X-axis
-      final int nextX = lastX == 0 ? minX : pwr2SeriesNext(xPPO, lastX);
+      final int nextX = lastX == 0 ? minX : (int)pwr2SeriesNext(xPPO, lastX);
       lastX = nextX;
       final int trials = getNumTrials(nextX);
       //configure();
