+++ src/main/java/org/apache/datasketches/characterization/req/ReqSketchAccuracyProfile.java
@@ -222,7 +222,7 @@ public class ReqSketchAccuracyProfile implements JobProfile {
 
       //go to next stream length
       if (useppo) {
-        streamLength = pwr2SeriesNext(ppo, streamLength);
+        streamLength = (int)pwr2SeriesNext(ppo, streamLength);
       } else {
         lgCurSL += lgDelta;
         streamLength = 1 << lgCurSL;
